+++ src/index.ts
@@ -1,22 +1,22 @@
 import * as util from './util';
 
-export * from './AccountTrackerController';
-export * from './AddressBookController';
-export * from './AssetsContractController';
-export * from './AssetsController';
-export * from './AssetsDetectionController';
-export * from './BaseController';
+export * from './assets/AccountTrackerController';
+export * from './user/AddressBookController';
+export * from './assets/AssetsContractController';
+export * from './assets/AssetsController';
+export * from './assets/AssetsDetectionController';
+export * from '../BaseController';
 export * from './ComposableController';
-export * from './CurrencyRateController';
-export * from './KeyringController';
-export * from './PersonalMessageManager';
-export * from './NetworkController';
-export * from './NetworkStatusController';
-export * from './PhishingController';
-export * from './PreferencesController';
-export * from './ShapeShiftController';
-export * from './TokenBalancesController';
-export * from './TokenRatesController';
-export * from './TransactionController';
-export * from './TypedMessageManager';
+export * from './assets/CurrencyRateController';
+export * from './keyring/KeyringController';
+export * from './message-manager/PersonalMessageManager';
+export * from './network/NetworkController';
+export * from './network/NetworkStatusController';
+export * from './third-party/PhishingController';
+export * from './user/PreferencesController';
+export * from './third-party/ShapeShiftController';
+export * from './assets/TokenBalancesController';
+export * from './assets/TokenRatesController';
+export * from './transaction/TransactionController';
+export * from './message-manager/TypedMessageManager';
 export { util };
