+++ src/message-manager/MessageManager.ts
@@ -1,5 +1,5 @@
 import { EventEmitter } from 'events';
-import BaseController, { BaseConfig, BaseState } from './BaseController';
+import BaseController, { BaseConfig, BaseState } from '../BaseController';
 
 /**
  * @type OriginalRequest
